#268 move stub-gen tests into separate file [skip ci]
